server/requirements.txt
asgiref==3.5.2
astroid==2.11.6
autopep8==1.6.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dill==0.3.5.1
dj-rest-auth==2.2.5
Django==4.0.5
django-allauth==0.51.0
django-configurations==2.3.2
django-cors-headers==3.13.0
django-rest-auth==0.9.5
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-stubs==1.7.0
drf-extra-fields==3.4.0
drf-writable-nested==0.7.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.3
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
types-Markdown==3.4.0
types-pytz==2022.1.1
types-PyYAML==6.0.9
types-requests==2.28.8
types-urllib3==1.26.22
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
